+++ src/main/kotlin/clean/Trashcan.kt
@@ -1,18 +1,14 @@
 package top.e404.eclean.clean
 
 import org.bukkit.event.EventHandler
+import org.bukkit.event.Listener
 import org.bukkit.event.inventory.InventoryCloseEvent
 import org.bukkit.inventory.ItemStack
-import top.e404.eclean.PL
 import top.e404.eclean.app.RuntimeServices
 import top.e404.eclean.config.Config
 import top.e404.eclean.menu.trashcan.TrashInfo
-import top.e404.eplugin.listener.EListener
 
-object Trashcan : EListener(PL) {
-    /**
-     * 垃圾桶中的物品及其数量, ItemStack的数量没有作用, 以value的数量为准
-     */
+object Trashcan : Listener {
     val trashData get() = RuntimeServices.trashcanRepository.trashData
 
     val trashValues get() = RuntimeServices.trashcanRepository.trashValues
@@ -21,9 +17,6 @@ object Trashcan : EListener(PL) {
         RuntimeServices.trashcanService.clearAll()
     }
 
-    /**
-     * 垃圾桶清空倒计时, 单位秒
-     */
     val countdown get() = RuntimeServices.trashcanService.countdown
 
     fun schedule() {
@@ -50,7 +43,6 @@ object Trashcan : EListener(PL) {
 
     @EventHandler
     fun InventoryCloseEvent.onEvent() {
-        // 菜单更新改为基于 MenuManager 的当前菜单快照，这里不再维护独立的打开集合
     }
 
     fun open(player: org.bukkit.entity.Player) {