+++ src/main/kotlin/util/Text.kt
@@ -21,3 +21,15 @@ fun String.placeholder(placeholder: Map<String, Any?>): String {
     for ((k, v) in placeholder.entries) s = s.replace("{$k}", v.toString())
     return s.color
 }
+
+fun Long.parseSecondAsDuration(): String {
+    if (this <= 0) return "0s"
+    val h = this / 3600
+    val m = this % 3600 / 60
+    val s = this % 60
+    return buildString {
+        if (h > 0) append("${h}h ")
+        if (m > 0) append("${m}min ")
+        append("${s}s")
+    }.trim()
+}